+++ src/main/java/com/zzheads/HomeAutomation/controller/EquipmentController.java
@@ -1,5 +1,8 @@
 package com.zzheads.HomeAutomation.controller;//
 
+import com.zzheads.HomeAutomation.exceptions.ApiErrorBadRequest;
+import com.zzheads.HomeAutomation.exceptions.ApiErrorNotFound;
+import com.zzheads.HomeAutomation.exceptions.DaoException;
 import com.zzheads.HomeAutomation.model.Equipment;
 import com.zzheads.HomeAutomation.model.Room;
 import com.zzheads.HomeAutomation.service.ControlService;
@@ -10,7 +13,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.Map;
 
 // HomeAutomation
@@ -22,29 +24,41 @@ public class EquipmentController {
     @Autowired EquipmentService equipmentService;
     @Autowired ControlService controlService;
 
+
+    private static final String expectedRequestFormat = "Can't make that request. Expected data format: {\"equipmentName\" : equipmentName}";
+    private boolean requestOk (Map request) {
+        return request.containsKey("equipmentName");
+    }
+
+
     @RequestMapping(value = "/room/{roomId}/equipment", method = RequestMethod.POST, produces = {"application/json"})
     @ResponseStatus(HttpStatus.OK)
-    public @ResponseBody Map addEquipment(@RequestBody Map<String, String> req, @PathVariable Long roomId) {
+    public @ResponseBody String addEquipment(@RequestBody Map<String, String> req, @PathVariable Long roomId) throws DaoException {
+        if (!requestOk(req)) throw new ApiErrorBadRequest(400, String.format("%s (%s)", expectedRequestFormat, Thread.currentThread().getStackTrace()[1].toString()));
         Equipment equipment = new Equipment(req);
         Room room = roomService.findById(roomId);
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (%s)", roomId, Thread.currentThread().getStackTrace()[1].toString()));
         equipment.setRoom(room);
         room.getEquipments().add(equipment);
         equipmentService.save(equipment);
         roomService.save(room);
-        return equipment.Json();
+        return equipment.toJson();
     }
 
     @RequestMapping(value = "/room/{roomId}/equipment", method = RequestMethod.GET, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody List getAllEquipment(@PathVariable Long roomId) {
-        return Equipment.Json(equipmentService.findByRoom(roomId));
+    public @ResponseBody String getAllEquipment(@PathVariable Long roomId) throws DaoException {
+        return Equipment.toJson(equipmentService.findByRoom(roomId));
     }
 
     @RequestMapping(value = "/room/{roomId}/equipment/{equipmentId}", method = RequestMethod.PUT, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map updateEquipment(@RequestBody Map<String, String> req, @PathVariable Long roomId, @PathVariable Long equipmentId) {
+    public @ResponseBody String updateEquipment(@RequestBody Map<String, String> req, @PathVariable Long roomId, @PathVariable Long equipmentId) throws DaoException {
+        if (!requestOk(req)) throw new ApiErrorBadRequest(400, String.format("%s (%s)", expectedRequestFormat, Thread.currentThread().getStackTrace()[1].toString()));
         Room room = roomService.findById(roomId);
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (updateEquipment method)", roomId));
         Equipment equipmentOld = equipmentService.findById(equipmentId);
+        if (equipmentOld == null) throw new ApiErrorNotFound(404, String.format("Can't find equipment with %d id. (%s)", equipmentId, Thread.currentThread().getStackTrace()[1].toString()));
         Equipment equipment = new Equipment(req);
         equipment.setId(equipmentId);
         equipment.setRoom(room);
@@ -52,28 +66,28 @@ public class EquipmentController {
         room.addEquipment(equipment);
         equipmentService.save(equipment);
         roomService.save(room);
-        return equipmentService.findById(equipmentId).Json();
+        return equipmentService.findById(equipmentId).toJson();
     }
 
     @RequestMapping(value = "/room/{roomId}/equipment/{equipmentId}", method = RequestMethod.GET, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map getEquipmentById(@PathVariable Long roomId, @PathVariable Long equipmentId) {
+    public @ResponseBody String getEquipmentById(@PathVariable Long roomId, @PathVariable Long equipmentId) throws DaoException {
         Room room = roomService.findById(roomId);
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (%s)", roomId, Thread.currentThread().getStackTrace()[1].toString()));
         Equipment equipment = equipmentService.findById(equipmentId);
-
-        return equipment.Json();
+        if (equipment == null) throw new ApiErrorNotFound(404, String.format("Can't find equipment with %d id. (%s)", equipmentId, Thread.currentThread().getStackTrace()[1].toString()));
+        return equipment.toJson();
     }
 
     @RequestMapping(value = "/room/{roomId}/equipment/{equipmentId}", method = RequestMethod.DELETE, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map deleteEquipmentById(@PathVariable Long roomId, @PathVariable Long equipmentId) {
+    public void deleteEquipmentById(@PathVariable Long roomId, @PathVariable Long equipmentId) throws DaoException {
         Room room = roomService.findById(roomId);
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (%s)", roomId, Thread.currentThread().getStackTrace()[1].toString()));
         Equipment equipment = equipmentService.findById(equipmentId);
+        if (equipment == null) throw new ApiErrorNotFound(404, String.format("Can't find equipment with %d id. (%s)", equipmentId, Thread.currentThread().getStackTrace()[1].toString()));
         room.removeEquipment(equipment);
         equipmentService.delete(equipment);
         roomService.save(room);
-
-        return null;
     }
-
 }
