+++ src/main/java/com/zzheads/HomeAutomation/controller/RoomController.java
@@ -1,5 +1,9 @@
 package com.zzheads.HomeAutomation.controller;//
 
+import com.google.gson.Gson;
+import com.zzheads.HomeAutomation.exceptions.ApiErrorBadRequest;
+import com.zzheads.HomeAutomation.exceptions.ApiErrorNotFound;
+import com.zzheads.HomeAutomation.exceptions.DaoException;
 import com.zzheads.HomeAutomation.model.Room;
 import com.zzheads.HomeAutomation.service.ControlService;
 import com.zzheads.HomeAutomation.service.EquipmentService;
@@ -9,8 +13,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -26,47 +28,51 @@ public class RoomController {
     @Autowired
     ControlService controlService;
 
-    public static final String BASE_URL ="http://localhost:8080/";
+    public static final String EXPECTED_REQUEST_FORMAT = "Can't make that request. Expected data format: {\"roomName\" : roomName, \"squareFootage\" : squareFootage}";
+    private boolean requestOk (Map request) {
+        return request.containsKey("roomName") && request.containsKey("squareFootage");
+    }
 
     @RequestMapping(value = "/room", method = RequestMethod.POST, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map addRoom(@RequestBody Map<String, String> req) {
+    public @ResponseBody String addRoom(@RequestBody Map<String, String> req) throws DaoException {
+        if (!requestOk(req)) throw new ApiErrorBadRequest(400, String.format("%s (%s)", EXPECTED_REQUEST_FORMAT, Thread.currentThread().getStackTrace()[1].toString()));
         Room room = new Room(req);
         roomService.save(room);
-        return room.Json();
+        return room.toJson();
     }
 
     @RequestMapping(value = "/room", method = RequestMethod.GET, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody List getAllRooms() {
+    public @ResponseBody String getAllRooms() throws DaoException {
+        Gson gson = new Gson();
         List<Room> rooms = roomService.findAll();
-
-        return Room.Json(rooms);
+        return Room.toJson(rooms);
     }
 
     @RequestMapping(value = "/room/{id}", method = RequestMethod.PUT, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map updateRoom(@RequestBody Map<String, String> req, @PathVariable Long id) {
+    public @ResponseBody String updateRoom(@RequestBody Map<String, String> req, @PathVariable Long id) throws DaoException {
+        if (requestOk(req)) throw new ApiErrorBadRequest(400, String.format("%s (%s)", EXPECTED_REQUEST_FORMAT, Thread.currentThread().getStackTrace()[1].toString()));
         Room room = new Room(req);
         room.setId(id);
         roomService.save(room);
-        return room.Json();
+        return room.toJson();
     }
 
     @RequestMapping(value = "/room/{id}", method = RequestMethod.GET, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map getRoomById(@PathVariable Long id) {
+    public @ResponseBody String getRoomById(@PathVariable Long id) throws DaoException {
         Room room = roomService.findById(id);
-
-        return room.Json();
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (%s)", id, Thread.currentThread().getStackTrace()[1].toString()));
+        return room.toJson();
     }
 
     @RequestMapping(value = "/room/{id}", method = RequestMethod.DELETE, produces = {"application/json"})
     @ResponseStatus (HttpStatus.OK)
-    public @ResponseBody Map deleteRoomById(@PathVariable Long id) {
+    public void deleteRoomById(@PathVariable Long id) throws DaoException {
         Room room = roomService.findById(id);
+        if (room == null) throw new ApiErrorNotFound(404, String.format("Can't find room with %d id. (%s)", id, Thread.currentThread().getStackTrace()[1].toString()));
         roomService.delete(room);
-
-        return null;
     }
 }
